% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peerRank.R
\name{pr}
\alias{pr}
\title{PeerRank score construction per Walsh (2014).}
\usage{
pr(A, alpha = 0.1, beta = 0.1, precision = 1e-10)
}
\arguments{
\item{A}{a square matrix of grades having been normalized in interval [0,1].}

\item{alpha}{an optional parameter. Must be in interval (0,1). Default = .10.}

\item{beta}{an optional parameter.  Must be in interval (0, 1], as must the
sum of alpha and beta.  The Generalized PeerRank rule degenerates to
the (non-generalized) PeerRank rule when beta is set to zero. Default
= .10.}
}
\description{
PeerRank score construction per Walsh (2014).
}
\references{
Walsh, T. 2014. The PeerRank method for peer assessment. In
            Proceedings of the 21st European Conference on Artificial
            Intelligence (ECAI), 909-914.
            doi: 10.3233/978-1-61499-419-0-909
}
